+++ tramtracker/asyncapi.py
@@ -73,14 +73,17 @@ class AsyncTramTrackerAPI(object):
         _logger.info("AsyncTramTrackerAPI instance deleted")
         return
 
-    async def call(self: Self, request: str) -> _responsemodel.TramTrackerResponse:
+    async def request(self: Self, path: str) -> _responsemodel.TramTrackerResponse:
         """Requests data from the TramTracker service and returns the response.
 
-        :param request: The request, which is appended to the base URL of the service
-        :return:        A :class:`list` or :class:`dict` of the response data, depending on the request
+        :param path: The request, which is appended to the base URL of the service
+        :return:     A :class:`list` or :class:`dict` of the response data, depending on the request
+
+        .. versionchanged:: 0.5.0
+            Renamed method from ``call``; now returns the whole server response including the enclosing object
         """
 
-        url = f"https://tramtracker.com.au/Controllers{request}"
+        url = f"https://tramtracker.com.au/Controllers{path}"
         _logger.debug("Entering rate limiter context manager")
         async with self._limiter:  # Rate limit requests
             _logger.debug("Requesting from: " + url)
@@ -110,7 +113,7 @@ class AsyncTramTrackerAPI(object):
         :return: A list detailing each route terminus
         """
 
-        response = cast(_responsemodel.DestinationsResponse, await self.call("/GetAllRoutes.ashx"))
+        response = cast(_responsemodel.DestinationsResponse, await self.request("/GetAllRoutes.ashx"))
         return [TramDestination(route_id=element["InternalRouteNo"],
                                 route_number=element["AlphaNumericRouteNo"] if element["AlphaNumericRouteNo"] is not None else str(element["RouteNo"]),
                                 up_direction=element["IsUpDestination"],
@@ -126,7 +129,7 @@ class AsyncTramTrackerAPI(object):
         :return:             A list of stops on the route
         """
 
-        response = cast(_responsemodel.StopsResponse, await self.call(f"/GetStopsByRouteAndDirection.ashx?r={route_id}&u={"true" if up_direction else "false"}"))
+        response = cast(_responsemodel.StopsResponse, await self.request(f"/GetStopsByRouteAndDirection.ashx?r={route_id}&u={"true" if up_direction else "false"}"))
         return [TramStop(stop_id=element["StopNo"],
                          stop_name=element["Description"],
                          stop_number=element["StopName"].split()[0],
@@ -145,7 +148,7 @@ class AsyncTramTrackerAPI(object):
         :return:        The stop details
         """
 
-        response = cast(_responsemodel.StopResponse, await self.call(f"/GetStopInformation.ashx?s={stop_id}"))
+        response = cast(_responsemodel.StopResponse, await self.request(f"/GetStopInformation.ashx?s={stop_id}"))
         element = response["ResponseObject"]
         return TramStop(stop_id=stop_id,
                         stop_name=element["StopName"],
@@ -165,7 +168,7 @@ class AsyncTramTrackerAPI(object):
         :return:        A list of route numbers
         """
 
-        response = cast(_responsemodel.RoutesResponse, await self.call(f"/GetPassingRoutes.ashx?s={stop_id}"))
+        response = cast(_responsemodel.RoutesResponse, await self.request(f"/GetPassingRoutes.ashx?s={stop_id}"))
         return [element["RouteNo"] for element in response["ResponseObject"]]
 
     async def next_trams(self: Self, stop_id: int, route_id: int | None = None, low_floor_tram: bool = False, as_of: datetime = datetime.now(tz=ZoneInfo("Australia/Melbourne"))) -> list[TramDeparture]:
@@ -181,7 +184,7 @@ class AsyncTramTrackerAPI(object):
             as_of = as_of.replace(tzinfo=TZ_MELBOURNE)
         as_of = as_of.astimezone(TZ_MELBOURNE)
         timestamp = round((as_of - EPOCH) / timedelta(milliseconds=1))
-        response = cast(_responsemodel.DeparturesResponse, await self.call(f"/GetNextPredictionsForStop.ashx?stopNo={stop_id}&routeNo={route_id if route_id is not None else 0}&isLowFloor={"true" if low_floor_tram else "false"}&ts={timestamp}"))
+        response = cast(_responsemodel.DeparturesResponse, await self.request(f"/GetNextPredictionsForStop.ashx?stopNo={stop_id}&routeNo={route_id if route_id is not None else 0}&isLowFloor={"true" if low_floor_tram else "false"}&ts={timestamp}"))
         return [TramDeparture(stop_id=stop_id,
                               trip_id=element["TripID"],
                               route_id=element["InternalRouteNo"],
@@ -216,7 +219,7 @@ class AsyncTramTrackerAPI(object):
         if as_of.tzinfo is None:
             as_of = as_of.replace(tzinfo=TZ_MELBOURNE)
         timestamp = round((as_of - datetime(1970, 1, 1, tzinfo=timezone.utc)) / timedelta(milliseconds=1))
-        response = cast(_responsemodel.ColourResponse, await self.call(f"/GetRouteColour.ashx?routeNo={route_id}&ts={timestamp}"))
+        response = cast(_responsemodel.ColourResponse, await self.request(f"/GetRouteColour.ashx?routeNo={route_id}&ts={timestamp}"))
         return "#" + response["responseObject"]["Colour"].lower()
 
     async def get_route_text_colour(self: Self, route_id: int, as_of: datetime = datetime.now(tz=TZ_MELBOURNE)) -> str:
@@ -229,5 +232,5 @@ class AsyncTramTrackerAPI(object):
         if as_of.tzinfo is None:
             as_of = as_of.replace(tzinfo=TZ_MELBOURNE)
         timestamp = round((as_of - datetime(1970, 1, 1, tzinfo=timezone.utc)) / timedelta(milliseconds=1))
-        response = cast(_responsemodel.ColourResponse, await self.call(f"/GetRouteTextColour.ashx?routeNo={route_id}&ts={timestamp}"))
+        response = cast(_responsemodel.ColourResponse, await self.request(f"/GetRouteTextColour.ashx?routeNo={route_id}&ts={timestamp}"))
         return "#" + response["responseObject"]["Colour"].lower()